Arguments & Main: Get user id

+++ src/arguments.rs
@@ -12,6 +12,9 @@ struct Cli {
 	#[clap(long, value_parser)]
 	add_playlist: Option<String>,
 	/// Update playlist stored in database
+	#[clap(long,action,value_parser)]
+	new_user_id: bool,
+	/// Update playlist stored in database
 	#[clap(short,long,action,value_parser)]
 	update: bool,
 	/// Delete a playlist
@@ -31,14 +34,16 @@ pub fn parse_args() -> Args{
 	let cli = Cli::parse();
 	let res;
 
-	if cli.update && cli.add_user == None && cli.add_playlist == None && cli.delete_playlist == None{
+	if cli.update && !cli.new_user_id && cli.add_user == None && cli.add_playlist == None && cli.delete_playlist == None{
 		res = Args::Update;
-	}else if !cli.update && cli.add_user != None && cli.add_playlist == None && cli.delete_playlist == None {
+	}else if !cli.update && !cli.new_user_id && cli.add_user != None && cli.add_playlist == None && cli.delete_playlist == None {
 		res = Args::NewUser(cli.add_user.unwrap());
-	}else if !cli.update && cli.add_user == None && cli.add_playlist != None && cli.delete_playlist == None {
+	}else if !cli.update && !cli.new_user_id && cli.add_user == None && cli.add_playlist != None && cli.delete_playlist == None {
 		res = Args::NewPlaylist(cli.add_playlist.unwrap());
-	}else if !cli.update && cli.add_user == None && cli.add_playlist == None && cli.delete_playlist != None {
+	}else if !cli.update && !cli.new_user_id && cli.add_user == None && cli.add_playlist == None && cli.delete_playlist != None {
 		res = Args::DeletePlaylist(cli.delete_playlist.unwrap());
+	}else if !cli.update && !cli.new_user_id && cli.add_user == None && cli.add_playlist == None && cli.delete_playlist == None {
+		res = Args::NewUserId;
 	}else{
 		let mut cmd = Cli::command();
 		cmd.error(
